implementing "other" for surveys
